LayoutTests/imported/w3c/web-platform-tests/trusted-types/worker-constructor.https-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'require-trusted-types-for'.


PASS Create Worker via ScriptTestUrl
PASS Create SharedWorker via ScriptTestUrl
PASS Create ServiceWorker via ScriptTestUrl
FAIL Block Worker creation via string assert_throws_js: function "() => new Worker(test_url)" did not throw
FAIL Block SharedWorker creation via string assert_throws_js: function "() => new SharedWorker(test_url)" did not throw
FAIL Block ServiceWorker creation via String assert_unreached: Should have rejected: undefined Reached unreachable code
PASS Setup default policy.
PASS Create Worker via string with default policy.
PASS Create SharedWorker via string with default policy.
FAIL Create ServiceWorker via string with default policy. promise_test: Unhandled rejection with value: object "TypeError: Script https://localhost:9443/trusted-types/support/WorkerGlobalScope-importScripts.potato.js load failed"
